requirements.txt
customtkinter==5.2.2
tkinterdnd2==0.3.0
pypdf==4.2.0
pdfplumber==0.11.0
pypdfium2==4.30.0
reportlab==4.2.0
pdf2docx==0.5.8
pdf2image==1.17.0
pytesseract==0.3.10
Pillow>=12.0.0
pikepdf==9.0.0
pyhanko==0.21.0
pyhanko-certvalidator==0.26.3
openpyxl==3.1.2
pandas==2.2.2
cryptography==42.0.7
requests==2.32.0
packaging==24.0
psutil==5.9.8
